feat: add validation to delete vehicle service

+++ server/src/services/deleteVehicle.ts
@@ -1,15 +1,24 @@
+import mongoose from "mongoose";
 import { Vehicle } from "../model/Vehicle";
 
 export const deleteVehicle = async (id: string) => {
-  try {
-    await Vehicle.findById(id);
+    try {
+        // Validate ObjectId
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            throw new Error('Invalid vehicle ID.');
+        }
 
-    await Vehicle.findByIdAndDelete(id);
+        // Check if vehicle exists
+        const existingVehicle = await Vehicle.findById(id);
+        if (!existingVehicle) {
+            throw new Error('Vehicle not found.');
+        }
 
-    return {
-      message: "Vehicle deleted successfully.",
-    };
-  } catch (error: any) {
-    throw new Error(error.message);
-  }
+        // Delete vehicle
+        await Vehicle.findByIdAndDelete(id);
+        return { message: 'Vehicle deleted successfully.' };
+
+    } catch (error: any) {
+        throw new Error(error.message);
+    }
 };
